[domain] move S3ObjectsData moved into module
