+++ domain/src/main/scala/net/kemitix/s3thorp/domain/S3ObjectsData.scala
@@ -1,6 +1,4 @@
-package net.kemitix.s3thorp.awssdk
-
-import net.kemitix.s3thorp.domain.{HashModified, KeyModified, MD5Hash, RemoteKey}
+package net.kemitix.s3thorp.domain
 
 /**
   * A list of objects and their MD5 hash values.
