+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3Client.scala
@@ -6,7 +6,7 @@ import com.amazonaws.services.s3.transfer.{TransferManager, TransferManagerBuild
 import com.github.j5ik2o.reactive.aws.s3.S3AsyncClient
 import com.github.j5ik2o.reactive.aws.s3.cats.S3CatsIOClient
 import net.kemitix.s3thorp._
-import net.kemitix.s3thorp.domain.{Bucket, HashModified, KeyModified, LocalFile, MD5Hash, RemoteKey}
+import net.kemitix.s3thorp.domain.{Bucket, HashModified, KeyModified, LocalFile, MD5Hash, RemoteKey, S3ObjectsData}
 
 trait S3Client {
